Sleazy vanilla-JS version of form
Adds a form with a numeric input and a submit-button to check that
number against the static fleet-number object. Moves the fleet-number
object from a reusable JSON file to a JS module for ease of use in this
sleazy context; might have to move it back to make it portable if we
take this elsewhere.

To deploy for local testing, run `python -m http.server` in this
directory.